Add TLS support to the memcached exporter
This is based on the changes in `gomemcache` PR
https://github.com/grobie/gomemcache/pull/2.

The original non-TLS behaviour is unchanged, however when `--tls.enable`
is given, the net connection is created by the `crypto/tls` module
instead of the `net` module.

The PR follows a similar setup to the TLS code in `amtool` and left
cert/key/ca/servername/insecure-skip-verify configurable. The
`ServerName` defaults to the provided address which seems a sensible
default. During testing, verification of the server certificates was
expecting an IP SAN even when a hostname is provided as the connection
address, hence the default.

Signed-off-by: Nick Rhodes <nrhodes91@gmail.com>