Fix unix domain socket connections.
The TLS connection code assumes the connection address will contain a port, which is not true for unix domain sockets. This change moves the host:port splitting code so that it is only run if TLS connections are enabled.

While it is possible TLS may be enabled on a unix socket, in that case the -memcached.tls.server-name flag will need to be set, which will bypass the call to SplitHostPort.

Signed-off-by: Gabe Van Engel <gabe@schizoid.net>